+++ ui/src/loudml/components/OneClickML.js
@@ -14,6 +14,7 @@ import {createHook} from 'src/loudml/utils/hook'
 import {
     normalizeInterval,
     normalizeFeatureDefault,
+    normalizeSpan,
 } from 'src/loudml/utils/model'
 import {
     createModel,
@@ -236,6 +237,7 @@ class OneClickML extends Component {
             max_evals: 10,
             name: this.name,
             interval: normalizeInterval(time),
+            span: normalizeSpan(time),
             default_datasource: datasource,
             bucket_interval: time,
             features: fields.map(